--- Assets/Scripts/UI.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 33e3c6b044404ec44b48c609fc0aec0f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
