--- Assets/Scripts/UI/Components/Specific_Components.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 210fe7c08a8553d49a71857b601a9881
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
