--- Assets/Scripts/UI/Components/ToggleGroup.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: aea92016926b29a4898918be8332e531
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
