+++ Assets/_Application/_Twin/UI/AdjustHeightOnTextChange.cs
@@ -1,9 +1,7 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using TMPro;
 
-namespace Netherlands3D.Twin
+namespace Netherlands3D.Twin.UI
 {
     public class AdjustHeightOnTextChange : MonoBehaviour
     {
