+++ Assets/_Application/_Twin/UI/ToggleGroupItem.cs
@@ -1,9 +1,7 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 
-namespace Netherlands3D.Twin
+namespace Netherlands3D.Twin.UI
 {
     public class ToggleGroupItem : MonoBehaviour
     {
