+++ Assets/_Application/_Twin/UI/Tooltips.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6a2309341a89b4a4a9cc9407d985a9ac
+guid: efbecf14d53f747e8ba68ad5a075f993
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
